% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.Code.R
\name{as.data.frame.Code}
\alias{as.data.frame.Code}
\title{Title}
\usage{
\method{as.data.frame}{Code}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Code object to convert}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column names}

\item{...}{additional arguments to be passed to or from methods}
}
\value{
data.frame
}
\description{
Title
}
\examples{
data(RS.data)
rs = RS.data
newcode = create.code(name = "Data", expressions = c("number","data"), excerpts = rs$text)
as.data.frame(newcode)

}
