% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.CodeSet.R
\name{as.data.frame.CodeSet}
\alias{as.data.frame.CodeSet}
\title{Title}
\usage{
\method{as.data.frame}{CodeSet}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{CodeSet to convert}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column names}

\item{...}{additional arguments to be passed to or from methods}
}
\value{
data.frame
}
\description{
Title
}
\examples{
data(RS.data)
rs = RS.data
newcode = create.code(name = "Data", expressions = c("number","data"), 
    excerpts = rs$text)
code.set = code.set("Demo RS CodeSet", "CodeSet made for the demo", 
    excerpts = rs$text, codes = c(newcode))
as.data.frame(code.set)

}
