% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeSet.R
\name{code.set}
\alias{code.set}
\title{Create CodeSet}
\usage{
code.set(title = "", description = "", excerpts = c(), codes = c())
}
\arguments{
\item{title}{Title for the CodeSet}

\item{description}{Description of the CodeSet}

\item{excerpts}{Set of excerpts to use with the CodeSet}

\item{codes}{Set of codes to attach to the CodeSet}
}
\value{
CodeSet object
}
\description{
Create a new CodeSet object
}
\examples{
data(RS.data)
rs = RS.data
code.set = code.set("Demo RS CodeSet", "CodeSet made for the demo", excerpts = rs$text, codes = c())

}
