% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{create.code}
\alias{create.code}
\title{Create a code}
\usage{
create.code(name = "NewCode", definition = "", excerpts = NULL,
  type = "Regex", ...)
}
\arguments{
\item{name}{Name of the code}

\item{definition}{Definition of the Code}

\item{excerpts}{Character vectore of excerpts to use for Coding}

\item{type}{Character string representing the type of code (Default: "Regex")}

\item{...}{Adiditional parameters}
}
\value{
Code object
}
\description{
Create a code
}
\examples{
data(RS.data)
rs = RS.data

# Generate a Code
newcode = create.code(name = "Data", expressions = c("number","data"), excerpts = rs$text)


}
