% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differences.R
\name{differences}
\alias{differences}
\title{Find Differences}
\usage{
differences(code = NULL)
}
\arguments{
\item{code}{Code object to search for differences}
}
\value{
logical vector representing indices that are coded differently
}
\description{
Find rows that differ within a data.frame or two vectors
}
\details{
Find rows that differ within a data.frame or two vectors
}
