% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegexCode.R
\name{expression.match}
\alias{expression.match}
\title{Expression Matching}
\usage{
expression.match(excerpts, expressions, names = list(NULL, "V1"))
}
\arguments{
\item{excerpts}{Character vector to match against}

\item{expressions}{Character vector of expressions}

\item{names}{Character vector to use for dimension names}
}
\value{
Matrix representing matched expressions
}
\description{
Match a set of text excerpts against a set of regular expressions
}
