% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHandSetIndices.R
\name{getHandSetIndices}
\alias{getHandSetIndices}
\title{Handset indices}
\usage{
getHandSetIndices(set, handSetLength = 20, handSetBaserate = 0.2)
}
\arguments{
\item{set}{[TBD]}

\item{handSetLength}{[TBD]}

\item{handSetBaserate}{[TBD]}
}
\description{
Handset indices
}
