% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHandSetIndices.R
\name{getHandSetIndices2}
\alias{getHandSetIndices2}
\title{Get indices to code}
\usage{
getHandSetIndices2(code, handSetLength = 20, handSetBaserate = 0.2,
  unseen = F, this.set = NULL)
}
\arguments{
\item{code}{Code object}

\item{handSetLength}{Number of excerpts to put into the test set}

\item{handSetBaserate}{Minimum number of positives that should be in the test set}

\item{unseen}{[TBD]}

\item{this.set}{[TBD]}
}
\value{
Code object with an updated test set and computer set
}
\description{
Get indices to code
}
