% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{old_test}
\alias{old_test}
\title{Calculate statistics}
\usage{
old_test(code, kappaThreshold = 0.65, baserateInflation = 0.2,
  type = c("training", "test"))
}
\arguments{
\item{code}{Code object to test}

\item{kappaThreshold}{Threshold used for calculating rhoR::rho}

\item{baserateInflation}{inflation rate to use when sampling handsets}

\item{type}{vector indicating which stats should be calculated}
}
\value{
Code object with updated statistics property
}
\description{
Run tests (kappa, rho) on the given Code
}
