% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary&print.R
\name{print.summary.Code}
\alias{print.summary.Code}
\title{Print a Code summary}
\usage{
\method{print}{summary.Code}(x, ...)
}
\arguments{
\item{x}{list from summary()}

\item{...}{Additional parameters}
}
\value{
Prints code summary
}
\description{
Print a Code summary
}
\examples{
data(RS.data)
rs = RS.data
newcode = create.code(name = "Data", 
    expressions = c("number","data"), excerpts = rs$text)
summary(newcode)

}
