% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary&print.R
\name{print.summary.CodeSet}
\alias{print.summary.CodeSet}
\title{Print the summary of a CodeSet}
\usage{
\method{print}{summary.CodeSet}(x, ...)
}
\arguments{
\item{x}{Summary of a CodeSet}

\item{...}{Additional parameters}
}
\value{
prints summary
}
\description{
Print the summary of a CodeSet
}
\examples{
data(RS.data)
rs = RS.data

newcode = create.code(name = "Data", 
    expressions = c("number","data"), excerpts = rs$text)
code.set = code.set("Demo RS CodeSet", "CodeSet made for the demo", 
    excerpts = rs$text, codes = c(newcode))
summary(code.set)

}
