% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary&print.R
\name{print.summary.Test}
\alias{print.summary.Test}
\title{Print a Test summary}
\usage{
\method{print}{summary.Test}(x, ...)
}
\arguments{
\item{x}{list from summary()}

\item{...}{Additional parameters}
}
\value{
prints summary
}
\description{
Print a Test summary
}
\examples{
data(RS.data)
rs = RS.data
newcode = create.code( name = "Data", 
    expressions = c("number","data"), excerpts = rs$text)
#newcode = handcode(code = newcode, excerpts = rs$text, n = 4)
newcode = test(code = newcode, kappaThreshold = 0.65)
summary(newcode$statistics)
}
