% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary&print.R
\name{print.summary.TestList}
\alias{print.summary.TestList}
\title{Print a TestList summary}
\usage{
\method{print}{summary.TestList}(x, ...)
}
\arguments{
\item{x}{list from summary()}

\item{...}{Additional parameters}
}
\value{
prints summary
}
\description{
Print a TestList summary
}
\examples{
data(RS.data)
rs = RS.data
newcode <- create.code("Data", expressions = c("number","data"), excerpts = rs$text)
newcode <- handcode(newcode, this.set = 10:15, results = 0)
newcode = test(code = newcode, kappa_threshold = 0.65)
summary(newcode$statistics)

}
