% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegexCode.R
\docType{data}
\name{RegexCode}
\alias{RegexCode}
\title{RegexCode}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RegexCode
}
\value{
RegexCode object
}
\description{
Creates an object for Regular Expression coding.  No need to call this
directly, create.code is a nice wrapper around this and any other types of Codes
}
\section{Fields}{

\describe{
\item{\code{name}}{Name of the Code}

\item{\code{definition}}{Definition of the Code}

\item{\code{excerpts}}{Character vector of text excerpts to code}

\item{\code{...}}{Additional parameters not specific to a RegexCode}

\item{\code{expressions}}{Character vector of regular expressions}
}}

\examples{
data(RS.data)
rs = RS.data

# Generate a Code
newcode = RegexCode$new(name = "New Code", definition = "Some definition", 
    excerpts = rs$text, expressions = c("number","data"))

}
\keyword{datasets}
