% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve}
\alias{resolve}
\title{Resolve differences}
\usage{
resolve(code = NULL, trainingSet = NULL, computerSet = NULL,
  expressions = NULL, excerpts = NULL, ignored = NULL)
}
\arguments{
\item{code}{Code to resolve coding differences}

\item{trainingSet}{Optionally provide a trainingSet, default: code$trainingSet}

\item{computerSet}{Optionally provide a computerSet, default: code$computerSet}

\item{expressions}{Optionally provide a set of expressions, default: code$expressions}

\item{excerpts}{Optionally provide a set of excerpts, default: code$excerpts}

\item{ignored}{Optionally proivde a set of excerpts to ignore during the resolve cycle loop}
}
\description{
Resolve differing results
}
