% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary&print.R
\name{summary.Code}
\alias{summary.Code}
\title{Obtain summary of a Code object}
\usage{
\method{summary}{Code}(object, ...)
}
\arguments{
\item{object}{Code to summarize}

\item{...}{Additional parameters}
}
\value{
List of Code summary
}
\description{
Obtain summary of a Code object
}
\examples{
data(RS.data)
rs = RS.data
newcode = create.code(name = "Data", 
    expressions = c("number","data"), excerpts = rs$text)
summary(newcode)

}
