% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary&print.R
\name{summary.CodeSet}
\alias{summary.CodeSet}
\title{Obtain a summary of the CodeSet}
\usage{
\method{summary}{CodeSet}(object, ...)
}
\arguments{
\item{object}{CodeSet object}

\item{...}{Additional parameters}
}
\value{
list containing description and Code summaries
}
\description{
Obtain a summary of the CodeSet
}
\examples{
data(RS.data)
rs = RS.data

newcode = create.code(name = "Data", 
    expressions = c("number","data"), excerpts = rs$text)
code.set = code.set("Demo RS CodeSet", "CodeSet made for the demo", 
    excerpts = rs$text, codes = c(newcode))
summary(code.set)

}
