% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary&print.R
\name{summary.Test}
\alias{summary.Test}
\title{Obtain a summary of a Code's test results}
\usage{
\method{summary}{Test}(object, ...)
}
\arguments{
\item{object}{Test object of Code}

\item{...}{Additional parameters}
}
\value{
list of Test summary
}
\description{
Obtain a summary of a Code's test results
}
\examples{
data(RS.data)
rs = RS.data
newcode = create.code(name = "Data", 
    expressions = c("number","data"), excerpts = rs$text)
#newcode = handcode(code = newcode, excerpts = rs$text, n = 4)
newcode = test(code = newcode, kappaThreshold = 0.65)
summary(newcode$statistics)
}
