% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{test}
\alias{test}
\title{Title}
\usage{
test(code, kappa_threshold = 0.65, baserate_inflation = 0.2, ...)
}
\arguments{
\item{code}{[TBD]}

\item{kappa_threshold}{[TBD]}

\item{baserate_inflation}{[TBD]}

\item{...}{[TBD]}
}
\value{
code object
}
\description{
Title
}
