% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocode.R
\name{autocode}
\alias{autocode}
\title{Match a list of expressions against some set of excerpts}
\usage{
autocode(x = NULL, expressions = NULL, excerpts = NULL,
  simplify = T, mode = "all")
}
\arguments{
\item{x}{Object to autocode. Either a Code or CodeSet}

\item{expressions}{Expressions to use for coding (optional)}

\item{excerpts}{Excerpts to code}

\item{simplify}{If TRUE, returns a data.frame, else returns a Code or CodeSet object}

\item{mode}{Either all, training, or test representing the set of excerpts that should be recoded in the computerSet}
}
\value{
data.frame of is simplify = T (default), otherwise the Code or CodeSet object with updated computerSets
}
\description{
Autocodes all codes provided, either directly with code or as part of a provided codeset
}
