% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differences.R
\name{differences}
\alias{differences}
\title{Find Differences}
\usage{
differences(code = NULL, wh = "trainingSet", to = "computerSet")
}
\arguments{
\item{code}{Code object to search for differences}

\item{wh}{Set to use as the base comparison}

\item{to}{Set to compare wh to}
}
\value{
logical vector representing indices that are coded differently

vector of indices representing differences
}
\description{
Find rows that differ within a data.frame or two vectors
}
\details{
Find rows that differ within a data.frame or two vectors
}
