% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHandSetIndices.R
\name{getHandSetIndices}
\alias{getHandSetIndices}
\title{Handset indices}
\usage{
getHandSetIndices(codeToUse, handSetLength = 20, handSetBaserate = 0.2,
  unseen = F)
}
\arguments{
\item{codeToUse}{[TBD]}

\item{handSetLength}{[TBD]}

\item{handSetBaserate}{[TBD]}

\item{unseen}{[TBD]}
}
\description{
Handset indices
}
