% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary&print.R
\name{summary.TestList}
\alias{summary.TestList}
\title{Obtain a summary of a Code's test results}
\usage{
\method{summary}{TestList}(object, ...)
}
\arguments{
\item{object}{TestList object of Code}

\item{...}{Additional parameters}
}
\value{
list of Test summary
}
\description{
Obtain a summary of a Code's test results
}
\examples{
data(RS.data)
rs = RS.data
newcode = create.code(name = "Data", 
    expressions = c("number","data"), excerpts = rs$text)
newcode <- handcode(newcode, this.set = 10:15, results = 0)
newcode = test(code = newcode, kappa_threshold = 0.65)
summary(newcode$statistics)
}
