% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{test}
\alias{test}
\title{Calculate statistics}
\usage{
test(code, kappaThreshold = 0.65)
}
\arguments{
\item{code}{Code object to test}

\item{kappaThreshold}{Threshold used for calculating rhoR::rho}
}
\value{
Code object with updated statistics property
}
\description{
Run tests (kappa, rho) on the given Code
}
